ai-faq/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
openai
